+++ package.json
@@ -4,7 +4,7 @@
   "description": "An open source logical way to keep context with information.",
   "main": "index.js",
   "scripts": {
-    "test": "./node_modules/.bin/jasmine-node test",
+    "test": "./node_modules/.bin/mocha ./test/**/*[sS]pec.js",
     "karma": "./node_modules/.bin/karma start"
   },
   "repository": {
@@ -27,17 +27,17 @@
   "homepage": "https://github.com/LamdaTech/Open-Graph-Book#readme",
   "devDependencies": {
     "babel-preset-es2015": "^6.14.0",
-    "jasmine": "^2.5.0",
-    "jasmine-core": "^2.5.0",
+    "chai": "^3.5.0",
     "karma": "^1.2.0",
+    "karma-chai": "^0.1.0",
     "karma-chrome-launcher": "^2.0.0",
     "karma-firefox-launcher": "^1.0.0",
-    "karma-jasmine": "^1.0.2",
+    "karma-mocha": "^1.1.1",
     "karma-phantomjs-launcher": "^1.0.2",
-    "koa": "^1.2.2"
+    "koa": "^1.2.2",
+    "mocha": "^3.0.2"
   },
   "dependencies": {
-    "jasmine-node": "^1.14.5",
     "request": "^2.74.0"
   }
 }
