--- src/app/app.js
@@ -1,11 +1,0 @@
-var koa = require('koa');
-var app = koa();
-
-console.log("Koa listening on localhost:3000")
-
-app.use(function *(){
-  this.body = 'Hello World';
-});
-
-app.listen(3000);
-
