--- src/app/helloworld.js
@@ -1,3 +1,0 @@
-exports.world = function() {
-  return 'Hello World!';
-}
