--- test/jasmine.json
@@ -1,9 +1,0 @@
-{
-  "spec_dir": "spec",
-  "spec_files": [
-    "spec/*[sS]pec.js"
-  ],
-  "src": ["../src/app/*.js"],
-  "stopSpecOnExpectationFailure": false,
-  "random": false
-}
